+++ packages/webview_flutter/webview_flutter_wkwebview/darwin/webview_flutter_wkwebview/Sources/webview_flutter_wkwebview/ErrorProxyAPIDelegate.swift
@@ -9,6 +9,13 @@ import Foundation
 /// This class may handle instantiating native object instances that are attached to a Dart instance
 /// or handle method calls on the associated native class or an instance of that class.
 class ErrorProxyAPIDelegate: PigeonApiDelegateNSError {
+  func pigeonDefaultConstructor(
+    pigeonApi: PigeonApiNSError, code: Int64, domain: String, userInfo: [String: Any?]
+  ) throws -> NSError {
+    return NSError(
+      domain: domain, code: Int(code), userInfo: userInfo.compactMapValues { $0 })
+  }
+
   func code(pigeonApi: PigeonApiNSError, pigeonInstance: NSError) throws -> Int64 {
     return Int64(pigeonInstance.code)
   }